Move grid benchmarks to benchmark executable
